move parse_library to core, checkout v_to
